server/requirements.txt
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
Flask==2.0.2
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pycparser==2.20
python-dotenv==0.19.1
six==1.16.0
Werkzeug==2.0.2
